Fix django-pipeline incompatibility
We can't use partial filenames in the {% static %} tag when using
django-pipeline, as it immediately tries to source the file on disk
and throws an error when it can't be found.
Instead, pass the STATIC_URL setting through to the template and use this.

This is a partial revert of 2b35e7b2a302c8396177d3f44cb998d2b8d715de.